moving parseLegacyItemToSandBoxItem to utils;
